requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.19
Werkzeug==2.3.6
Flask-MySQLdb==1.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
